+++ src/main/java/com/devHyun/TodayTrip01/domain/course/CourseRepository.java
@@ -1,0 +1,6 @@
+package com.devHyun.TodayTrip01.domain.course;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CourseRepository extends JpaRepository<Course, Long> {
+}
