--- src/main/java/com/devHyun/TodayTrip01/domain/spot/Address.java
@@ -1,9 +1,0 @@
-package com.devHyun.TodayTrip01.domain.spot;
-
-public class Address {
-
-    private String city;
-    private String street;
-
-
-}
