+++ src/main/java/com/devHyun/TodayTrip01/domain/spot/Spot.java
@@ -1,12 +1,10 @@
 package com.devHyun.TodayTrip01.domain.spot;
 
 
-import com.devHyun.TodayTrip01.domain.trip.Course;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import org.springframework.beans.factory.annotation.Value;
 
 import javax.persistence.*;
 
@@ -25,7 +23,7 @@ public class Spot {
     private String title;               // 관광지 이름
 
     private String detail;              // 관광지 상세정보(overview)
-    private String firstimage;          // 관광지 이미지
+    private String image;          // 관광지 이미지
 
     private String address;             // 주소
     private String homepage;            // 홈페이지
