--- src/main/java/com/devHyun/TodayTrip01/domain/spot/SpotList.java
@@ -1,11 +1,0 @@
-package com.devHyun.TodayTrip01.domain.spot;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class SpotList {
-
-    public static List<Spot> spotList = new ArrayList<>();
-
-
-}
