--- src/main/java/com/devHyun/TodayTrip01/domain/trip/TripRepository.java
@@ -1,6 +1,0 @@
-package com.devHyun.TodayTrip01.domain.trip;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TripRepository extends JpaRepository<Trip, Long> {
-}
