--- src/main/resources/application.properties
@@ -1,1 +1,0 @@
-server.post=80
